moves map under data-tool layout
