+++ frontend/src/containers/data-tool/content/map/sync-settings.ts
